Move fetchDataset to helper file
