+++ test/ivr/benchmark-source.test.js
@@ -25,10 +25,10 @@ describe('souce loads records', () => {
     const source = new BenchmarkSource()
     const records = await source.loadAll()
     expect(records.length).toBe(5)
-    expect(records[0].utterance).toBe('six')
+    expect(records[0].utterance).toBe('seven')
     expect(records[0].meta.platform).toBe('twilio-autopilot')
     expect(records[3].meta.number).toBe('1123456789')
-    expect(records[0].meta.clipUrl).toBe('common_voice_en_21953345.mp3')
+    expect(records[0].meta.clipUrl).toBe('common_voice_en_22497731.mp3')
   })
 
   test('source handles amazon connect platform correctly', async () => {
@@ -44,9 +44,9 @@ describe('souce loads records', () => {
     const source = new BenchmarkSource()
     const records = await source.loadAll()
     expect(records.length).toBe(5)
-    expect(records[1].utterance).toBe('yes')
+    expect(records[1].utterance).toBe('eigth')
     expect(records[1].meta.platform).toBe('amazon-connect')
     expect(records[1].meta.number).toBe('1198765432')
-    expect(records[1].meta.clipUrl).toBe('common_voice_en_22214552.mp3')
+    expect(records[1].meta.clipUrl).toBe('common_voice_en_22497742.mp3')
   })
 })
